+++ src/components/TransactionsTable.tsx
@@ -7,16 +7,11 @@ import TableBody from "@mui/material/TableBody";
 import TableCell from "@mui/material/TableCell";
 import TableHead from "@mui/material/TableHead";
 import TableRow from "@mui/material/TableRow";
-import Box from "@mui/material/Box";
 import Typography from "@mui/material/Typography";
 import {useNavigate} from "react-router-dom";
-import Popover from "@mui/material/Popover";
-import {IconButton} from "@mui/material";
-import {grey} from "../themes/colors/aptosColorPalette";
-import ChevronRightRoundedIcon from "@mui/icons-material/ChevronRightRounded";
-import ChevronLeftRoundedIcon from "@mui/icons-material/ChevronLeftRounded";
-
-import Button from "@mui/material/Button";
+import GeneralTableRow from "../components/GeneralTableRow";
+import GeneralTableHeaderCell from "../components/GeneralTableHeaderCell";
+import HashButton from "./HashButton";
 
 import {Types} from "aptos";
 import {
@@ -26,7 +21,6 @@ import {
   renderTransactionType,
 } from "../pages/Transactions/helpers";
 import {assertNever} from "../utils";
-import {truncateAddress} from "../pages/utils";
 
 type TransactionCellProps = {
   transaction: Types.OnChainTransaction;
@@ -82,115 +76,9 @@ function TransactionGasCell({transaction}: TransactionCellProps) {
 }
 
 function TransactionHashCell({transaction}: TransactionCellProps) {
-  const [anchorEl, setAnchorEl] = React.useState<HTMLButtonElement | null>(
-    null,
-  );
-
-  const hashExpand = (event: React.MouseEvent<HTMLButtonElement>) => {
-    event.stopPropagation();
-    setAnchorEl(event.currentTarget);
-  };
-
-  const hashCollapse = (event: React.MouseEvent<HTMLButtonElement>) => {
-    event.stopPropagation();
-    setAnchorEl(null);
-  };
-
-  const open = Boolean(anchorEl);
-  const id = open ? "simple-popover" : undefined;
-  const theme = useTheme();
-
   return (
     <TableCell>
-      <Box>
-        <Button
-          sx={{
-            textTransform: "none",
-            backgroundColor: `${
-              theme.palette.mode === "dark" ? grey[700] : grey[100]
-            }`,
-            display: "flex",
-            borderRadius: 1,
-            color: "inherit",
-            padding: "0.15rem 0.5rem 0.15rem 1rem",
-            "&:hover": {
-              backgroundColor: `${
-                theme.palette.mode === "dark" ? grey[700] : grey[100]
-              }`,
-            },
-          }}
-          aria-describedby={id}
-          onClick={hashExpand}
-          variant="contained"
-          endIcon={<ChevronRightRoundedIcon sx={{opacity: "0.75", m: 0}} />}
-        >
-          {truncateAddress(transaction.hash)}
-        </Button>
-
-        <Popover
-          onClick={(e) => {
-            e.stopPropagation();
-          }}
-          sx={{
-            overflow: "scroll",
-            ".MuiPaper-root": {boxShadow: "none"},
-            "&.MuiModal-root .MuiBackdrop-root": {
-              transition: "none!important",
-              backgroundColor: `${
-                theme.palette.mode === "dark"
-                  ? "rgba(18,22,21,0.5)"
-                  : "rgba(255,255,255,0.5)"
-              }`,
-            },
-          }}
-          id={id}
-          open={open}
-          anchorEl={anchorEl}
-          onClose={hashCollapse}
-          anchorOrigin={{
-            vertical: "center",
-            horizontal: "left",
-          }}
-          transformOrigin={{
-            vertical: "center",
-            horizontal: "left",
-          }}
-        >
-          <Typography
-            sx={{
-              display: "flex",
-              alignItems: "center",
-              backgroundColor: `${
-                theme.palette.mode === "dark" ? grey[700] : grey[100]
-              }`,
-              px: 2,
-              py: "0.15rem",
-              fontSize: "14px",
-              overflow: "scroll",
-              scrollbarWidth: "none",
-              "&::-webkit-scrollbar": {
-                display: "none",
-              },
-            }}
-          >
-            {transaction.hash}
-            <IconButton
-              aria-label="collapse hash"
-              onClick={hashCollapse}
-              sx={{
-                ml: 1,
-                mr: 0,
-                p: 0,
-                "&.MuiButtonBase-root:hover": {
-                  bgcolor: "transparent",
-                },
-              }}
-            >
-              <ChevronLeftRoundedIcon sx={{opacity: "0.5"}} />
-            </IconButton>
-          </Typography>
-        </Popover>
-      </Box>
+      <HashButton hash={transaction.hash} />
     </TableCell>
   );
 }
@@ -223,39 +111,17 @@ type TransactionRowProps = {
 function TransactionRow({transaction, columns}: TransactionRowProps) {
   const navigate = useNavigate();
 
-  const rowClick = (event: React.MouseEvent<unknown>) => {
+  const rowClick = () => {
     navigate(`/txn/${transaction.version}`);
   };
 
-  const theme = useTheme();
-
   return (
-    <TableRow
-      onClick={(event) => rowClick(event)}
-      sx={{
-        cursor: "pointer",
-        userSelect: "none",
-        backgroundColor: `${
-          theme.palette.mode === "dark" ? grey[800] : grey[50]
-        }`,
-        "&:hover:not(:active)": {
-          filter: `${
-            theme.palette.mode === "dark"
-              ? "brightness(0.9)"
-              : "brightness(0.99)"
-          }`,
-        },
-        "&:active": {
-          background: theme.palette.neutralShade.main,
-          transform: "translate(0,0.1rem)",
-        },
-      }}
-    >
+    <GeneralTableRow onClick={rowClick}>
       {columns.map((column) => {
         const Cell = TransactionCells[column];
         return <Cell key={column} transaction={transaction} />;
       })}
-    </TableRow>
+    </GeneralTableRow>
   );
 }
 
@@ -265,83 +131,36 @@ type TransactionHeaderCellProps = {
 
 function TransactionHeaderCell({column}: TransactionHeaderCellProps) {
   const theme = useTheme();
-  const tableCellBackgroundColor = "transparent";
-  const tableCellTextColor = theme.palette.text.secondary;
 
   switch (column) {
     case "version":
-      return (
-        <TableCell
-          sx={{
-            textAlign: "left",
-            background: `${tableCellBackgroundColor}`,
-            color: `${tableCellTextColor}`,
-          }}
-        >
-          <Typography variant="subtitle1">Version</Typography>
-        </TableCell>
-      );
+      return <GeneralTableHeaderCell header="Version" />;
     case "type":
-      return (
-        <TableCell
-          sx={{
-            textAlign: "left",
-            background: `${tableCellBackgroundColor}`,
-            color: `${tableCellTextColor}`,
-          }}
-        >
-          <Typography variant="subtitle1">Type</Typography>
-        </TableCell>
-      );
+      return <GeneralTableHeaderCell header="Type" />;
     case "hash":
-      return (
-        <TableCell
-          sx={{
-            textAlign: "left",
-            background: `${tableCellBackgroundColor}`,
-            color: `${tableCellTextColor}`,
-          }}
-        >
-          <Typography variant="subtitle1">Hash</Typography>
-        </TableCell>
-      );
+      return <GeneralTableHeaderCell header="Hash" />;
     case "status":
       return (
-        <TableCell
+        <GeneralTableHeaderCell
+          header="Status"
           sx={{
-            textAlign: "left",
-            background: `${tableCellBackgroundColor}`,
-            color: `${tableCellTextColor}`,
             borderRadius: `${theme.shape.borderRadius}px 0 0 ${theme.shape.borderRadius}px`,
           }}
-        >
-          <Typography variant="subtitle1">Status</Typography>
-        </TableCell>
+        />
       );
     case "gas":
       return (
-        <TableCell
+        <GeneralTableHeaderCell
+          header="Gas"
+          textAlignRight={true}
           sx={{
-            textAlign: "right",
-            background: `${tableCellBackgroundColor}`,
-            color: `${tableCellTextColor}`,
             borderRadius: `0 ${theme.shape.borderRadius}px ${theme.shape.borderRadius}px 0`,
           }}
-        >
-          <Typography variant="subtitle1">Gas</Typography>
-        </TableCell>
+        />
       );
     case "timestamp":
       return (
-        <TableCell
-          sx={{
-            textAlign: "right",
-            background: `${tableCellBackgroundColor}`,
-            color: `${tableCellTextColor}`,
-          }}
-        >
-          <Typography variant="subtitle1">Timestamp</Typography>
-        </TableCell>
+        <GeneralTableHeaderCell header="Timestamp" textAlignRight={true} />
       );
     default:
       return assertNever(column);
